Extract notifier behaviour
