+++ config/dev.exs
@@ -42,8 +42,8 @@ config :tilex, Tilex.Repo,
 
 config :tilex, :page_size, 50
 config :tilex, :auth_controller, AuthController
-config :tilex, :slack_notifier, Tilex.Notifications.Slack
-config :tilex, :twitter_notifier, Tilex.Notifications.Twitter
+config :tilex, :slack_notifier, Tilex.Notifications.Notifiers.Slack
+config :tilex, :twitter_notifier, Tilex.Notifications.Notifiers.Twitter
 config :tilex, :organization_name, System.get_env("ORGANIZATION_NAME")
 config :tilex, :canonical_domain, System.get_env("CANONICAL_DOMAIN")
 config :tilex, :cors_origin, "http://localhost:3000"
