+++ config/prod.exs
@@ -75,8 +75,8 @@ end
 
 config :tilex, :page_size, 50
 config :tilex, :auth_controller, AuthController
-config :tilex, :slack_notifier, Tilex.Notifications.Slack
-config :tilex, :twitter_notifier, Tilex.Notifications.Twitter
+config :tilex, :slack_notifier, Tilex.Notifications.Notifiers.Slack
+config :tilex, :twitter_notifier, Tilex.Notifications.Notifiers.Twitter
 config :tilex, :organization_name, System.get_env("ORGANIZATION_NAME")
 config :tilex, :ga_identifier, System.get_env("GA_IDENTIFIER")
 config :tilex, :canonical_domain, System.get_env("CANONICAL_DOMAIN")
